requirements.txt
cassandra-driver==3.29.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.4.0
geomet==0.2.1.post1
idna==3.7
py4j==0.10.9.7
pyspark==3.5.0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
